1049 no auto in public interface (#1050)
* no auto in public interface

* no auto in public interface second part

* fixed a return value

* more improvements

* improved doc for some utils, and use ranges::count

* more comments

+++ include/wex/ui/notebook.h
@@ -138,7 +138,7 @@ template <class T> bool wex::notebook::for_each(int id)
   const bool keys_empty(m_keys.empty());
 
   // The page should be an int (no), otherwise page >= 0 never fails!
-  for (int page = GetPageCount() - 1; page >= 0; page--)
+  for (int page = (int)GetPageCount() - 1; page >= 0; page--)
   {
     switch (T* win = reinterpret_cast<T*>(GetPage(page)); id)
     {
